+++ src/Passes/Source/Apps/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(qat Qat/Qat.cpp)
+add_executable(qat Qat/Qat.cpp Qat/LlvmAnalysis.cpp)
 
 target_link_libraries(qat ${llvm_libs})
 target_link_libraries(qat ExpandStaticAllocation QubitAllocationAnalysis TransformationRule Rules AllocationManager Commandline Profiles)
