+++ src/Passes/Source/Rules/Notation/Call.ipp
@@ -1,0 +1,41 @@
+#pragma once
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+#include "Llvm/Llvm.hpp"
+#include "Rules/Notation/Notation.hpp"
+#include "Rules/Operands/Any.hpp"
+#include "Rules/Operands/Call.hpp"
+#include "Rules/Operands/Instruction.hpp"
+#include "Rules/ReplacementRule.hpp"
+
+#include <unordered_map>
+#include <vector>
+
+namespace microsoft {
+namespace quantum {
+namespace notation {
+
+using OperandPrototypePtr = std::shared_ptr<OperandPrototype>;
+
+template <typename... Args>
+OperandPrototypePtr Call(std::string const &name, Args... args)
+{
+  OperandPrototypePtr              ret = std::make_shared<CallPattern>(name);
+  std::vector<OperandPrototypePtr> arguments{args...};
+
+  // Adding arguments to matching
+  for (auto &a : arguments)
+  {
+    ret->addChild(a);
+  }
+
+  // Function name is kept in the last operand
+  ret->addChild(std::make_shared<AnyPattern>());
+
+  return ret;
+}
+
+}  // namespace notation
+}  // namespace quantum
+}  // namespace microsoft
